verificacion de roles

+++ TP/client/src/components/Home/LastVacancies.vue
@@ -25,12 +25,14 @@
           <p v-if="vacante.vacantes_disponibles > 1"><i class="fas fa-exclamation-circle"></i>&nbsp;¡Quedan solo {{ vacante.vacantes_disponibles }} vacantes!</p>
           <p v-if="vacante.vacantes_disponibles === 1"><i class="fas fa-exclamation-circle"></i>&nbsp;¡Última vacante disponible!</p>
         </div>
-        <button @click="postularme(vacante.id)" class="btn btn-primary" v-if="!vacante.usuarioPostulado">
-          Postularme
-        </button>
-        <button @click="darmeDeBaja(vacante.id)" class="btn btn-danger" v-if="vacante.usuarioPostulado">
-          Darme de baja
-        </button>
+        <div v-if="!authenticated || (authenticated && isUsuario)">
+          <button @click="postularme(vacante.id)" class="btn btn-primary" v-if="!vacante.usuarioPostulado">
+            Postularme
+          </button>
+          <button @click="darmeDeBaja(vacante.id)" class="btn btn-danger" v-if="vacante.usuarioPostulado">
+            Darme de baja
+          </button>
+        </div>
       </div>
     </div>
     <Popup dataTarget="loginPostulacionPopup" title="Iniciar Sesión" :showButtons="false">
@@ -41,7 +43,7 @@
 
 <script>
 import axios from 'axios'
-import { mapActions } from 'vuex'
+import { mapGetters, mapActions } from 'vuex'
 import EventBus from '../../event-bus'
 export default {
   components: {
@@ -54,23 +56,31 @@ export default {
       id_llamado: null
     }
   },
+  computed: {
+    ...mapGetters({
+      authenticated: 'authenticated',
+      isUsuario: 'isUsuario',
+    })
+  },
   methods: {
     ...mapActions({
         logOut: 'logOut'
       }),
 
     async buscarPostulacionesDelUsuario() {
-      try {
-        let res = await axios.get('/postulaciones/buscarPostulacionesDelUsuario',
-        {
-          headers: {
-            Authorization: 'Bearer ' + this.$store.getters.user.api_token
-          }
-        });
+      if (this.$store.getters.authenticated && this.$store.getters.isUsuario) {
+        try {
+          let res = await axios.get('/postulaciones/buscarPostulacionesDelUsuario',
+          {
+            headers: {
+              Authorization: 'Bearer ' + this.$store.getters.user.api_token
+            }
+          });
 
-        this.postulacionesDelUsuario = res.data;
-      } catch (err) {
-        console.log(err.response.data.error);
+          this.postulacionesDelUsuario = res.data;
+        } catch (err) {
+          console.log(err.response.data.error);
+        }
       }
     },
 
@@ -105,7 +115,7 @@ export default {
       this.postulacionesDelUsuario = [];
       this.vacantes = [];
 
-      if (this.$store.getters.authenticated) {
+      if (this.$store.getters.authenticated && this.$store.getters.isUsuario) {
         await this.buscarPostulacionesDelUsuario();
       }
       await this.buscarVacantes();
@@ -113,21 +123,23 @@ export default {
 
     async postularme(id_llamado) {
       if (this.$store.getters.authenticated) {
-        try {
-          await axios.post('/postulaciones/agregarPostulacionDelUsuario',
-          {
-            id_llamado,
-            curriculum_vitae: 'curriculum.jpg'
-          },
-          {
-            headers: {
-              Authorization: 'Bearer ' + this.$store.getters.user.api_token
-            }
-          });
+        if (this.$store.getters.isUsuario) {
+          try {
+            await axios.post('/postulaciones/agregarPostulacionDelUsuario',
+            {
+              id_llamado,
+              curriculum_vitae: 'curriculum.jpg'
+            },
+            {
+              headers: {
+                Authorization: 'Bearer ' + this.$store.getters.user.api_token
+              }
+            });
 
-          this.actualizarVacantes();
-        } catch (err) {
-          console.log(err.response.data.error);
+            this.actualizarVacantes();
+          } catch (err) {
+            console.log(err.response.data.error);
+          }
         }
       } else {
         this.id_llamado = id_llamado;
@@ -137,17 +149,19 @@ export default {
 
     async darmeDeBaja(id_llamado) {
       if (this.$store.getters.authenticated) {
-        try {
-          await axios.delete('/postulaciones/eliminarPostulacionDelUsuario/' + id_llamado,
-          {
-            headers: {
-              Authorization: 'Bearer ' + this.$store.getters.user.api_token
-            }
-          });
-          
-          this.actualizarVacantes();
-        } catch (err) {
-          console.log(err.response.data.error);
+        if (this.$store.getters.isUsuario) {
+          try {
+            await axios.delete('/postulaciones/eliminarPostulacionDelUsuario/' + id_llamado,
+            {
+              headers: {
+                Authorization: 'Bearer ' + this.$store.getters.user.api_token
+              }
+            });
+            
+            this.actualizarVacantes();
+          } catch (err) {
+            console.log(err.response.data.error);
+          }
         }
       } else {
         this.logOut();
